feat: organized testing suite into :basic and :full
