+++ src/index.ts
@@ -2304,6 +2304,7 @@ const peanut = {
 	createClaimXChainPayload,
 	detectContractVersionFromTxReceipt,
 	estimateGasLimit,
+	formatNumberAvoidScientific,
 	generateKeysFromString,
 	getAllCreatedLinksForAddress,
 	getAllDepositsForSigner,
@@ -2339,6 +2340,7 @@ const peanut = {
 	solidityHashBytesEIP191,
 	supportsEIP1559,
 	toggleVerbose,
+	trim_decimal_overflow,
 	verifySignature,
 	VERSION,
 	CHAIN_DETAILS,
@@ -2367,6 +2369,7 @@ export {
 	createClaimXChainPayload,
 	detectContractVersionFromTxReceipt,
 	estimateGasLimit,
+	formatNumberAvoidScientific,
 	generateKeysFromString,
 	getAllCreatedLinksForAddress,
 	getAllDepositsForSigner,
@@ -2400,6 +2403,7 @@ export {
 	solidityHashBytesEIP191,
 	supportsEIP1559,
 	toggleVerbose,
+	trim_decimal_overflow,
 	verifySignature,
 	VERSION,
 	CHAIN_DETAILS,
