+++ test/basic/__hashesAndSigs.ts
