+++ test/basic/createClaimXChainPayload.test.ts
@@ -1,6 +1,6 @@
 import { ethers } from 'ethersv5'
 import { expect, it, describe, jest } from '@jest/globals'
-import peanut from '../src/index'
+import peanut from '../../src/index'
 
 describe('createClaimXChainPayload tests', function () {
 	afterEach(() => {
