+++ test/basic/getAllCreatedLinksForAddress.test.ts
@@ -1,5 +1,5 @@
 // Generated by CodiumAI
-import peanut from '../src/index' // import directly from source code
+import peanut from '../../src/index' // import directly from source code
 import { ethers } from 'ethersv5' // v5
 import dotenv from 'dotenv'
