+++ test/basic/getLinksFromTx.test.ts
@@ -1,4 +1,4 @@
-import peanut from '../src/index' // import directly from source code
+import peanut from '../../src/index' // import directly from source code
 import { ethers } from 'ethersv5' // v5
 import { expect, it, describe } from '@jest/globals'
 import dotenv from 'dotenv'
