+++ test/basic/prepareTxs.test.ts
@@ -1,8 +1,8 @@
-import peanut from '../src/index' // import directly from source code
+import peanut from '../../src/index' // import directly from source code
 import { ethers } from 'ethersv5' // v5
 import { expect, describe, it } from '@jest/globals'
 import dotenv from 'dotenv'
-import * as interfaces from '../src/consts/interfaces.consts'
+import * as interfaces from '../../src/consts/interfaces.consts'
 dotenv.config()
 
 const TEST_WALLET_PRIVATE_KEY = process.env.TEST_WALLET_PRIVATE_KEY ?? ''
