+++ test/basic/version.test.ts
@@ -1,11 +1,11 @@
-import peanut from '../src/index'
+import peanut from '../../src/index'
 import { dirname, join } from 'path'
 import { readFileSync } from 'fs'
 import dotenv from 'dotenv'
 dotenv.config()
 const directName = dirname(__filename)
 
-const packagePath = join(directName, '../package.json')
+const packagePath = join(directName, '../../package.json')
 console.log(packagePath)
 const packageJson = JSON.parse(readFileSync(packagePath, 'utf8'))
 const version = packageJson.version
@@ -16,7 +16,6 @@ describe('Peanut SDK', function () {
 
 		it('should return the current version', function () {
 			expect(peanut.VERSION).toBe(version)
-			expect(peanut.version).toBe(version)
 		})
 	})
 })
